hw2_affine/requirements.txt
# Panorama Affine Transform Dependencies

# Core libraries

opencv-python>=4.5.0
opencv-contrib-python>=4.5.0  # 包含 xfeatures2d (SURF)
numpy>=1.19.0
PyYAML>=5.3.0

# Optional utilities

pytest>=6.0.0      # 單元測試

# 額外支援 (視需求啟用)

# matplotlib>=3.3.0  # 若需繪圖展示
